requirements.txt
requests
pdfplumber
pydantic
sqlalchemy
psycopg2-binary
google-genai
pandas
numpy
openpyxl==3.1.2
streamlit